More SortOptions to arrow_schema
